split TC.lib into interface and implementation
