+++ .shared/commons.gradle
@@ -78,59 +78,3 @@ subprojects { subProject ->
 }
 
 
-// -----------------------------------------------------------------------------
-// Eclipse
-// -----------------------------------------------------------------------------
-
-// This is a cool example of the power of gradle.  Eclipse can't handle hierarchical projects.
-// so this is a clever way to use gradle to filter out which projects we want to apply the eclipse
-// plugin too. Suggestion comes from
-// http://gradle.1045684.n5.nabble.com/Best-practice-to-exclude-directories-from-generating-project-classpath-files-td4288890.html
-def leafProjects = subprojects.findAll {
-    it.childProjects.size() == 0
-}
-configure(leafProjects) {
-    apply plugin: 'eclipse'
-    eclipse {
-        //want parts of paths in resulting .classpath file to be replaced by variables
-        //pathVariables 'GRADLE_USER_HOME': gradle.gradleUserHomeDir
-
-		classpath {
-			//default settings for downloading sources and Javadoc:
-			downloadSources = true
-			downloadJavadoc = false
-		}
-    }
-// experimental: (use different folders for test and main)
-//	eclipse.classpath {
-//		defaultOutputDir = file('bin')
-//		file.withXml { n ->
-//		n.asNode().classpathentry.findAll { it.@path.startsWith('src/test')}
-//				.each { it.@output = 'bin-test' }
-//		}
-//	}
-
-	eclipse.classpath.file {
-		def projectEntries
-	    beforeMerged { classpath ->
-			//classpath.entries.findAll { entry -> println entry }
-			//classpath.entries.removeAll { entry -> entry.kind == 'lib' || entry.kind == 'var' }
-			//classpath.entries.removeAll { entry -> entry.path.startsWith("/") }
-			projectEntries = classpath.entries.findAll { entry -> entry.path.startsWith("/") }
-			//classpath.entries.findAll { entry -> println entry }
-		}
-		whenMerged { classpath ->
-			//classpath.entries.findAll { entry -> entry.kind == 'lib' }*.exported = false
-			//classpath.entries.findAll { ['con', 'lib'].contains(it.kind) }*.exported = false
-
-			//keep only those that are not already added by gradle
-			projectEntries.removeAll { projectEntry -> classpath.entries.find{ entry -> entry.path == projectEntry.path }}
-
-			//try to remove lib dependencies that match additionally included projects (experimental)
-			classpath.entries.removeAll { entry -> projectEntries.find{ projectEntry -> entry.path.contains(projectEntry.path) }}
-
-			// re-add additionally included projects
-			classpath.entries += projectEntries
-		}
-	}
-}
