+++ TC.exec/TC.exec.gradle
@@ -99,7 +99,7 @@ dependencies {
 	// local project dependencies
     implementation project(':Command')
     implementation project(':MessagingHelpers')
-	implementation project(':TC.lib')
+	implementation project(':TC.lib.if')
     
 	// external project dependencies
 	implementation libraries.slf4j_api
