+++ TC.lib.if/src/main/java/de/fraunhofer/iosb/tc_lib_if/IVCT_BaseModelIf.java
@@ -1,0 +1,33 @@
+/*
+Copyright 2021, Reinhard Herzog (Fraunhofer IOSB)
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package de.fraunhofer.iosb.tc_lib_if;
+
+import org.slf4j.Logger;
+
+
+/**
+ * @author Reinhard Herzog (Fraunhofer IOSB)
+ */
+public interface IVCT_BaseModelIf  {
+
+    public void startup();
+    public void shutdown();
+    public void setSettingsDesignator(String settingsDesignator);
+    public void setFederationName(String federationName);
+	public String getSutFederateName();
+	public void setSutFederateName(String sutFederateName);
+}
