+++ settings.gradle
@@ -2,6 +2,7 @@ include 'LogSink'
 include 'LogSink.webApp'
 include 'IEEE1516e'
 include 'TC.exec'
+include 'TC.lib.if'
 include 'TC.lib'
 include 'TC.webApp'
 include 'MessagingHelpers'
